Refactor Loginpage
